add sort userlist

+++ src/app/admin/components/list-users/list-users.component.ts
@@ -1,7 +1,7 @@
 import { AfterViewInit, Component, Input, OnChanges, OnInit, SimpleChanges, ViewChild } from '@angular/core';
 import { User } from '../../../auth/models/user.model';
 import { MatPaginator } from '@angular/material/paginator';
-import { MatSort } from '@angular/material/sort';
+import { MatSort, Sort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
 import { WalletService } from '../../../wallet/services/wallet.service';
 import { MatDialog, MatDialogRef } from '@angular/material/dialog';
@@ -12,6 +12,7 @@ import { ConfirmOperationDialogComponent } from '../../../shared/dialog/confirm-
 import { ToastService } from '../../../shared/services/toast.service';
 import { Wallet } from '../../../models/wallet/wallet.models';
 import { ShowWalletDialogComponent } from '../show-wallet-dialog/show-wallet-dialog.component';
+import { LiveAnnouncer } from '@angular/cdk/a11y';
 
 @Component({
   selector: 'app-list-users',
@@ -21,17 +22,21 @@ import { ShowWalletDialogComponent } from '../show-wallet-dialog/show-wallet-dia
 
 export class ListUsersComponent implements OnInit, AfterViewInit {
 
+
   @ViewChild (MatPaginator) paginator: MatPaginator | null = null;
   @ViewChild (MatSort) sort: MatSort | null = null;
 
   public displayedColumns: string[] = ['index', 'id', 'name', 'email', 'username', 'password', 'document', 'admin', 'edit', 'delete', 'wallet' ];
   public dataSource: MatTableDataSource<User> = new MatTableDataSource<User>();
 
+  public passwordVisibility: Map<string, boolean> = new Map();
+
   constructor (
     private walletService: WalletService,
     private authService: AuthService,
     private dialog: MatDialog,
     private toastService: ToastService,
+    private liveAnnouncer: LiveAnnouncer,
   ) {}
 
   ngOnInit(): void {
@@ -41,13 +46,21 @@ export class ListUsersComponent implements OnInit, AfterViewInit {
   public ngAfterViewInit(): void {
     this.dataSource.paginator = this.paginator;
     this.dataSource.sort = this.sort;
+
+    this.sort!.active = 'username';
+    this.sort!.direction = 'desc';
   }
 
-  public showWallet ( id: string ) : void {
+  togglePasswordVisibility(userId: string): void {
+    const isVisible = this.passwordVisibility.get(userId) || false;
+    this.passwordVisibility.set(userId, !isVisible);
+  }
 
-    if ( !id ) return;
+  public showWallet ( user: User ) : void {
 
-    this.walletService.getWalletByIdUser( id )
+    if ( user.admin ) return this.toastService.showError('Error', 'Los administradores no poseen billetera');
+
+    this.walletService.getWalletByIdUser( user.id )
     .pipe(
       tap( data => { if (!data) return this.toastService.showInfo('Info', 'El usuario no contiene billetera') }),
     )
@@ -71,7 +84,9 @@ export class ListUsersComponent implements OnInit, AfterViewInit {
   private getUsers () : void {
 
     this.authService.getAllUsers.subscribe( users => {
+      users.forEach( user => this.passwordVisibility.set(user.id, false));
       this.dataSource.data = users;
+
     });
 
   }
@@ -99,24 +114,28 @@ export class ListUsersComponent implements OnInit, AfterViewInit {
 
   public deleteUser (user : User) : void {
 
-    const dialogRef = this.dialog.open( ConfirmOperationDialogComponent );
+    if( user.admin ) return this.toastService.showError('Error','No es posible eliminar administradores');
 
+    const dialogRef = this.dialog.open( ConfirmOperationDialogComponent );
 
     dialogRef.afterClosed()
     .pipe(
       filter(confirm => !!confirm),
-      tap( () => { if (user.admin) return this.toastService.showInfo('Info','No es posible eliminar administradores') }),
       switchMap(() => this.walletService.deleteWalletByIdUser(user.id)),
       switchMap(() => this.authService.deleteUserById(user.id)),
-      tap(data => {
-        data
-          ? this.toastService.showSuccess('Éxito', 'Operación realizada con éxito!')
-          : this.toastService.showError('Error', 'No es posible eliminar administradores!');
-      }),
+      tap( () =>  this.toastService.showSuccess('Éxito', 'Operación realizada con éxito!')),
       switchMap(() => this.authService.getAllUsers)
     )
     .subscribe(users => this.dataSource.data = users);
 
+  }
+
+  public announceSortChange(sortState: Sort) {
+    if (sortState.direction) {
+      this.liveAnnouncer.announce(`Sorted ${sortState.direction}ending`);
+    } else {
+      this.liveAnnouncer.announce('Sorting cleared');
+    }
   }
